// src/app/shared/enum/movement-type.ts
export enum MovementType {
  SALE = 'SALE',
  LOAD = 'LOAD',
  UNLOAD = 'UNLOAD',
  INV = 'INV',
  PURCH = 'PURCH',
  RPURC = 'RPURC',
  RSALE = 'RSALE',
  RESE = 'RESE',
  ORDE = 'ORDE',
  MOVTO = 'MOVTO',
  MOVFROM = 'MOVFROM',
  LORDER = 'LORDER',
  URESER = 'URESER',
  RETURN = 'RETURN',
  CREATE = 'CREATE',
  UPDATE = 'UPDATE',
}